requirements.txt
langchain-core
langchain-community
langchain
chromadb
sentence-transformers
pyyaml
requests
httpx
faiss-cpu